requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
black==22.8.0
cachetools==4.2.4
certifi==2024.6.2
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
datasets==2.4.0
dill==0.3.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
future==1.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
huggingface-hub==0.4.0
hydra-core==1.3.2
idna==3.7
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
Markdown==3.3.7
multidict==5.2.0
multiprocess==0.70.12.2
mypy==0.971
mypy-extensions==1.0.0
numpy==1.19.5
oauthlib==3.2.2
omegaconf==2.3.0
packaging==21.3
pandas==1.1.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.6
pyarrow==6.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.10
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.27.1
requests-oauthlib==2.0.0
responses==0.17.0
rsa==4.9
sacremoses==0.0.53
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.12.1
tomli==1.2.3
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchmetrics==0.8.2
torchvision==0.11.2+cu111
tqdm==4.64.1
transformers==4.18.0
typed-ast==1.5.5
typing_extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
xxhash==3.2.0
yarl==1.7.2
zipp==3.6.0
